+++ CHANGE_LOG.md
@@ -5,6 +5,33 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project/module adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
+## V2.1.0 - 24.04.2025
+
+### Changed
+ - State changed before calling "on_entry" callback
+ - In case state change in "on_entry" callback don't execute "on_activity" handle
+
+### Todo
+ - State transition allowance & checking
+
+---
+## V2.0.0 - 26.09.2024
+
+### Notice
+Complete module rework, not compatible with older version!
+
+### Added
+ - Ability to exchange generic data across FSM states within FSM module (Look at *fsm_get_data*, *fsm_set_data* API)
+ - Added actions for entry & exit states
+
+### Changed
+ - Remove configuration for max allowed states in FSM
+ - FSM configuration table format change
+
+### Todo
+ - State transition allowance & checking
+
+---
 ## V1.2.1 - 27.09.2023
 
 ### Fixing